requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
djangorestframework==3.15.2
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
